refactor: 특정 기능의 컴포넌트 app 하위 폴더로 이동 (#40)
# Conflicts:
#	src/app/(home)/_components/home-posts.tsx
#	src/app/page.tsx

# Conflicts:
#	src/app/(home)/_components/home-posts.tsx
