+++ src/app/(home)/_components/banner.tsx
