+++ src/app/auth/_components/welcome-modal.tsx
